requirements.txt
Flask==2.1.1
Flask-SQLAlchemy==2.5.1
pandas==1.3.3
